added link in description of UserInfoActivity
