+++ app/src/androidTest/java/com/freshdigitable/udonroad/util/UserUtil.java
@@ -16,6 +16,7 @@
 
 package com.freshdigitable.udonroad.util;
 
+import twitter4j.URLEntity;
 import twitter4j.User;
 
 import static org.mockito.Mockito.mock;
@@ -31,6 +32,8 @@ public class UserUtil {
     when(mock.getName()).thenReturn("Akihito Matsuda");
     when(mock.getId()).thenReturn(2000L);
     when(mock.getProfileBackgroundColor()).thenReturn("ffffff");
+    when(mock.getDescription()).thenReturn("user description is here.");
+    when(mock.getDescriptionURLEntities()).thenReturn(new URLEntity[0]);
     return mock;
   }
 }
