+++ app/src/main/java/com/freshdigitable/udonroad/SpanningInfo.java
@@ -18,20 +18,10 @@ package com.freshdigitable.udonroad;
 
 import android.support.annotation.Nullable;
 import android.text.style.ClickableSpan;
-import android.text.style.URLSpan;
-import android.view.View;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-
-import twitter4j.ExtendedMediaEntity;
-import twitter4j.Status;
-import twitter4j.URLEntity;
-import twitter4j.UserMentionEntity;
 
 /**
+ * SpanningInfo is information to create SpannableStringBuilder.
+ *
  * Created by akihit on 2016/08/20.
  */
 class SpanningInfo {
@@ -53,82 +43,5 @@ class SpanningInfo {
 
   boolean isReplacing() {
     return displayingText != null;
-  }
-
-  static List<SpanningInfo> create(Status bindingStatus) {
-    List<SpanningInfo> info = new ArrayList<>();
-    info.addAll(createURLSpanningInfo(bindingStatus));
-    info.addAll(createUserSpanningInfo(bindingStatus));
-    for (int i = info.size() - 1; i >= 0; i--) {
-      final SpanningInfo spanningInfo = info.get(i);
-      for (int j = 0; j < i; j++) {
-        if (spanningInfo.start == info.get(j).start) {
-          info.remove(spanningInfo);
-          break;
-        }
-      }
-    }
-    // to manipulate changing tweet text length, sort descending
-    Collections.sort(info, new Comparator<SpanningInfo>() {
-      @Override
-      public int compare(SpanningInfo l, SpanningInfo r) {
-        return r.start - l.start;
-      }
-    });
-    return info;
-  }
-
-  private static List<SpanningInfo> createURLSpanningInfo(Status bindingStatus) {
-    final String text = bindingStatus.getText();
-    final URLEntity[] urlEntities = bindingStatus.getURLEntities();
-    final String quotedStatusIdStr = Long.toString(bindingStatus.getQuotedStatusId());
-    final List<SpanningInfo> info = new ArrayList<>();
-    for (URLEntity u : urlEntities) {
-      int start = text.indexOf(u.getURL());
-      int end = start + u.getURL().length();
-      if (isInvalidRange(text, start, end)) {
-        continue;
-      }
-      if (bindingStatus.getQuotedStatus() != null
-          && u.getExpandedURL().contains(quotedStatusIdStr)) {
-        info.add(new SpanningInfo(null, start, end, ""));
-      }
-      info.add(new SpanningInfo(new URLSpan(u.getExpandedURL()), start, end, u.getDisplayURL()));
-    }
-    final ExtendedMediaEntity[] eme = bindingStatus.getExtendedMediaEntities();
-    for (ExtendedMediaEntity e : eme) {
-      int start = text.indexOf(e.getURL());
-      int end = start + e.getURL().length();
-      if (isInvalidRange(text, start, end)) {
-        continue;
-      }
-      info.add(new SpanningInfo(null, start, end, ""));
-    }
-    return info;
-  }
-
-  private static List<SpanningInfo> createUserSpanningInfo(Status bindingStatus) {
-    final String text = bindingStatus.getText();
-    final UserMentionEntity[] userMentionEntities = bindingStatus.getUserMentionEntities();
-    final List<SpanningInfo> info = new ArrayList<>();
-    for (UserMentionEntity u : userMentionEntities) {
-      final int start = text.indexOf("@" + u.getScreenName());
-      final int end = start + u.getScreenName().length() + 1;
-      if (isInvalidRange(text, start, end)) {
-        continue;
-      }
-      final long id = u.getId();
-      info.add(new SpanningInfo(new ClickableSpan() {
-        @Override
-        public void onClick(View view) {
-          UserInfoActivity.start(view.getContext(), id);
-        }
-      }, start, end, null));
-    }
-    return info;
-  }
-
-  private static boolean isInvalidRange(String text, int start, int end) {
-    return start < 0 || end > text.length() || start > text.length();
   }
 }
