+++ app/src/main/java/com/freshdigitable/udonroad/realmdata/URLEntityRealm.java
@@ -16,12 +16,17 @@
 
 package com.freshdigitable.udonroad.realmdata;
 
+import com.android.annotations.NonNull;
+
+import io.realm.RealmList;
 import io.realm.RealmObject;
 import io.realm.annotations.Ignore;
 import io.realm.annotations.PrimaryKey;
 import twitter4j.URLEntity;
 
 /**
+ * URLEntityRealm is realm object implementation for URLEntity of twitter4j.
+ *
  * Created by akihit on 2016/06/23.
  */
 public class URLEntityRealm extends RealmObject implements URLEntity {
@@ -44,6 +49,19 @@ public class URLEntityRealm extends RealmObject implements URLEntity {
     this.start = urlEntity.getStart();
     this.end = urlEntity.getEnd();
   }
+
+  @NonNull
+  static RealmList<URLEntityRealm> createList(URLEntity[] urlEntities) {
+    if (urlEntities == null) {
+      return new RealmList<>();
+    }
+    RealmList<URLEntityRealm> urlEntityRealms = new RealmList<>();
+    for (URLEntity u : urlEntities) {
+      urlEntityRealms.add(new URLEntityRealm(u));
+    }
+    return urlEntityRealms;
+  }
+
   @Override
   public String getText() {
     return url;
